+++ src/utils.rs
@@ -1,5 +1,10 @@
+use aes::cipher::{block_padding::Pkcs7, BlockDecryptMut, KeyIvInit};
+use anyhow::{anyhow, bail, Result};
+use base64::Engine;
 use kdam::term::Colorizer;
 use regex::Regex;
+
+type Aes128CbcDec = cbc::Decryptor<aes::Aes128>;
 
 pub(super) fn format_bytes(bytesval: usize, precision: usize) -> (String, String, String) {
     let mut val = bytesval as f32;
@@ -78,6 +83,51 @@ pub(super) fn scrape_playlist_msg(url: &str) -> String {
         "INPUT".colorize("bold green"),
         url,
     )
+}
+
+pub(super) fn decode_base64<T: AsRef<[u8]>>(input: T) -> Result<Vec<u8>> {
+    base64::engine::general_purpose::STANDARD
+        .decode(input)
+        .map_err(|x| x.into())
+}
+
+// pub(super) fn encode_base64<T: AsRef<[u8]>>(input: T) -> String {
+//     base64::engine::general_purpose::STANDARD.encode(input)
+// }
+
+pub(super) fn decrypt_aes_128_cbc(
+    input: &mut [u8],
+    key: &[u8],
+    iv: Option<&Vec<u8>>,
+) -> Result<Vec<u8>> {
+    let key_length = key.len();
+
+    if key_length != 16 {
+        bail!(
+            "invalid key size i.e. {} but expected size 16.",
+            key_length
+        );
+    }
+
+    let mut key_c = [0_u8; 16];
+    key_c.copy_from_slice(key);
+
+    let mut iv_c = [0_u8; 16];
+
+    if let Some(iv) = iv {
+        let iv_length = key.len();
+
+        if iv_length != 16 {
+            bail!("invalid iv size i.e. {} but expected size 16.", iv_length);
+        }
+
+        iv_c.copy_from_slice(iv);
+    }
+
+    Aes128CbcDec::new(&key_c.into(), &iv_c.into())
+        .decrypt_padded_mut::<Pkcs7>(input)
+        .map(|x| x.to_vec())
+        .map_err(|x| anyhow!("{}", x))
 }
 
 // fn find_ffmpeg() -> Option<String> {